Data.Binary.Builder.Internal

Description

A module containing semi-public Builder internals that exposes low level construction functions. Modules which extend the Builder system will need to use this module while ideally most users will be able to make do with the public interface modules.

Low-level construction of Builders

writeN :: Int -> (Ptr Word8 -> IO ()) -> Builder

Ensure that n bytes are available, and then use f to write exactly n bytes into memory.

writeAtMost :: Int -> (Ptr Word8 -> IO Int) -> Builder

Ensure that n bytes are available, and then use f to write at most n bytes into memory. f must return the actual number of bytes written.